requirements.txt
azure-identity==1.18.0
azure-mgmt-compute==33.0.0
azure-mgmt-network==27.0.0
azure-mgmt-subscription==3.1.1
requests==2.31.0
